Relayer: move bitcoin to ika2btc package (#192)
* move bitcoin package to ika2btc


* add missing package


* mispelling


* bump go-version to 1.24


---------
