+++ .github/workflows/lint.yml
@@ -22,7 +22,7 @@ jobs:
       - uses: actions/setup-go@v5
         if: env.GIT_DIFF
         with:
-          go-version: "1.23"
+          go-version: "1.24"
           cache: true
       - name: Install dependencies
         run: |
