+++ .github/workflows/tests.yml
@@ -49,7 +49,7 @@ jobs:
       - uses: actions/setup-go@v5
         if: env.GIT_DIFF
         with:
-          go-version: "1.23"
+          go-version: "1.24"
           cache: true
       - name: Test and Create Coverage Report
         if: env.GIT_DIFF
