+++ ika2btc/bitcoin/client_mock.go
