+++ ika2btc/bitcoin/errors.go
